requirements.txt
fastapi>=0.110.0
python-multipart>=0.0.9
supabase>=2.3.4
python-dotenv>=1.0.0
pydantic>=2.6.0
email-validator>=2.1.0.post1
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
requests>=2.31.0
openpyxl>=3.1.2
# Explicitly install numpy first-compatible with Python 3.13 wheels
numpy>=2.1.0
pandas>=2.2.2
websockets>=12.0
